flask_middleware/requirements.txt
flask
grpcio
grpcio-tools
flask[async]
pytest